feat: offline viewing of agent traces
